+++ weiss_core/tests/action_space_reachability_tests.rs
@@ -332,7 +332,7 @@ fn action_space_ids_are_reachable() {
                     card_id: CARD_CHAR,
                     instance_id: (idx + 1) as u32,
                     zone: ChoiceZone::Hand,
-                    index: Some(idx as u8),
+                    index: Some(idx as u16),
                     target_slot: None,
                 })
                 .collect();
